binder: add active thread tracking and deferred free
Co-developed with Joel Fernandez <joelaf@google.com>

The binder driver code assumes that it is always safe
to dereference pointers for proc, thread, node,
and ref objects since the global lock used to guarantee
that they would not be freed while another thread was
referencing them.

To preserve this assumption, the notion of "zombie" lists
is added to preserve objects that have been released. A
zombie object can no longer be found by binder threads,
but a thread that already has a pointer to a zombie object
can safely dereference it.

Where the driver used to kfree() an object, that object is
now placed on a zombie list and the corresponding binder_proc
is placed on the global "zombie_procs" list for cleanup.

The zombie objects are freed once we know that all the
binder threads that might have had a pointer to the
object have either exited the kernel or are waiting
and no longer assume the pointers are valid.

To do this, we assign a monotonically increasing seq number
to active threads via binder_get_thread() and add to
a list ordered by seq number. Nodes are removed via
binder_put_thread() signalling the thread is no longer active.

Likewise, when a binder proc is added to the zombie_procs
list, it is assigned a seq number.

Any proc in zombie_procs that has a lower seq number than
the oldest active thread can safely be cleaned up and the
zombie objects freed. This is checked when a binder thread
leaves binder_ioctl().

The fact that a binder_proc is on the zombie_procs list
means that there is one or more objects "owned" by that
proc (node, thread, ref) that need cleanup. The binder
proc itself will only be cleaned up and freed if it was
marked as a zombie too.

It has been pointed out that using object references could
have been used for this. That was the original plan, and
was implemented for binder_refs. However, we found that
for some of the other objects that are derefed via pointers
from the proc or node, this became very complex and error
prone and to be viable would require much more refactoring.
The zombie model preserved the assumptions about object
persistence without significant refactoring.

Bug: 33250092 32225111
Change-Id: I97fe66f6554ecf96050cbc74ec028363e988f75b
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>